+++ backend/app/__init__.py
@@ -1,7 +1,7 @@
 # backend/app/__init__.py
 from flask import Flask
 from config import Config
-from app.extensions import db, migrate, cors
+from app.extensions import db, migrate, cors, bcrypt
 
 def create_app(config_class=Config):
     app = Flask(__name__)
@@ -12,8 +12,12 @@ def create_app(config_class=Config):
         from app import models
     migrate.init_app(app, db)
     cors.init_app(app, origins=app.config["CORS_ORIGINS"])
+    bcrypt.init_app(app)
 
     from app.routes.health import health_bp
     app.register_blueprint(health_bp)
+
+    from app.routes.auth import auth_bp
+    app.register_blueprint(auth_bp)
 
     return app